+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- (browser) Add `Bugsnag-Uncompressed-Content-Length` header to trace payloads [#280](https://github.com/bugsnag/bugsnag-js-performance/pull/280)
+
+### Fixed
+
+- (browser) Do not retry delivery for oversized payloads when connection fails [#280](https://github.com/bugsnag/bugsnag-js-performance/pull/280)
+
 ## v1.0.0 (2023-07-27)
 
 ### Added